Move method for login
